Revert "Create Plugin: Add metadata file to configs (#441)"
This reverts commit d1c750d6931a4f27fc31644462243cf7ed6387aa.
