+++ packages/create-plugin/src/commands/version.command.ts
@@ -1,4 +1,5 @@
-import { getVersion } from '../utils/utils.version';
+import { resolve } from 'path';
+import { readFileSync } from 'fs';
 
 export const version = async () => {
   try {
@@ -7,4 +8,14 @@ export const version = async () => {
     console.error(error);
     process.exit(1);
   }
+};
+
+const getVersion = () => {
+  const packageJsonPath = resolve(__dirname, '..', '..', 'package.json');
+  const pkg = readFileSync(packageJsonPath, 'utf8');
+  const { version } = JSON.parse(pkg);
+  if (!version) {
+    throw `Could not find the version of sign-plugin`;
+  }
+  return version;
 };
