+++ packages/create-plugin/src/utils/utils.templates.ts
@@ -2,12 +2,12 @@ import glob from 'glob';
 import path from 'path';
 import fs from 'fs';
 import mkdirp from 'mkdirp';
-import { filterOutCommonFiles, getExportFileName, isFile, isFileStartingWith } from './utils.files';
+import { filterOutCommonFiles, isFile, isFileStartingWith } from './utils.files';
 import { renderHandlebarsTemplate } from './utils.handlebars';
 import { getPluginJson } from './utils.plugin';
 import { TEMPLATE_PATHS, EXPORT_PATH_PREFIX, EXTRA_TEMPLATE_VARIABLES } from '../constants';
 import { getPackageManagerWithFallback } from './utils.packageManager';
-import { getVersion } from './utils.version';
+import { getExportFileName } from '../utils/utils.files';
 
 /**
  *
@@ -77,6 +77,5 @@ export function getTemplateData() {
     pluginType: pluginJson.type,
     packageManagerName,
     packageManagerVersion,
-    createPluginVersion: getVersion(),
   };
 }
