--- packages/create-plugin/templates/common/.config/.create-plugin.json
@@ -1,3 +1,0 @@
-{
-  "version": "{{ createPluginVersion }}"
-}
